pyproject.toml
[project]
name = "weather-bot"
version = "0.1.0"
description = "Weather bot for Telegram"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.0,<4.0",
    "python-dotenv",
    "httpx",
    "fastapi",
    "uvicorn",
]

